Update to bullet-2.77.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@7579 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/bullet/src/LinearMath/btStackAlloc.h
@@ -23,13 +23,14 @@ Nov.2006
 #include "btScalar.h" //for btAssert
 #include "btAlignedAllocator.h"
 
+///The btBlock class is an internal structure for the btStackAlloc memory allocator.
 struct btBlock
 {
 	btBlock*			previous;
 	unsigned char*		address;
 };
 
-///StackAlloc provides some fast stack-based memory allocator (LIFO last-in first-out)
+///The StackAlloc class provides some fast stack-based memory allocator (LIFO last-in first-out)
 class btStackAlloc
 {
 public:
@@ -61,7 +62,7 @@ public:
 
 	int	getAvailableMemory() const
 	{
-		return totalsize - usedsize;
+		return static_cast<int>(totalsize - usedsize);
 	}
 
 	unsigned char*			allocate(unsigned int size)
