+++ src/tracks/bezier_curve.cpp
@@ -38,7 +38,8 @@ BezierCurve::BezierCurve(const XMLNode &node)
 Vec3 BezierCurve::getXYZ(float t) const
 {
     unsigned int i=int(t);   // FIXME: have to figure out which point we want here
-    if(i>=m_all_data.size()-1) return m_all_data[i].m_control_point;
+    if(i>=(unsigned int)m_all_data.size()-1) 
+        return m_all_data[i].m_control_point;
 
     const BezierData &p0 = m_all_data[i];
     const BezierData &p1 = m_all_data[i+1];
